Client::Doorkeeper -> Providers::Doorkeeper
